refactor: LineStatus enum을 domain 패키지에서 view 패키지로 이동
